+++ tests/shipping/Scheduled Delivery_Scheduled Pickup - Credit card - vtexgame1.test.js
@@ -1,0 +1,5 @@
+
+  import test from "./models/Scheduled Delivery_Scheduled Pickup - Credit card.model.js"
+
+  test("vtexgame1")
+